qa_project/requirements.txt
openai>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
datasets>=2.14.0
transformers>=4.35.0
rouge-score>=0.1.2
nltk>=3.8.1
colorama>=0.4.6
tabulate>=0.9.0
